refactor(frontend): extract hierarchy name cell renderer into its own module
Move renderNameCell (the hierarchy grid's name column: indentation, the
expand/collapse toggle, the name text, and the hover action overlay) out of
HierarchyColumns into hierarchyNameCell, carrying the EXPAND_ICON_SLOT_SIZE
constant with it. HierarchyColumns drops the ExpandMore/ChevronRight icons,
contextMenuActionsOverlaySx, IconButton/Tooltip, and the renderInlineActions
import it no longer needs; it shrinks from ~403 to ~293 lines. No behavior
change.
